fixed incorrect displaying of word + punctuation in Hindi
