+++ app/src/main/java/io/github/sspanak/tt9/ime/modes/ModeABC.java
@@ -10,7 +10,7 @@ import io.github.sspanak.tt9.languages.Language;
 import io.github.sspanak.tt9.languages.LanguageKind;
 import io.github.sspanak.tt9.languages.NaturalLanguage;
 import io.github.sspanak.tt9.preferences.settings.SettingsStore;
-import io.github.sspanak.tt9.util.Characters;
+import io.github.sspanak.tt9.util.chars.Characters;
 
 class ModeABC extends InputMode {
 	private final ArrayList<ArrayList<String>> KEY_CHARACTERS = new ArrayList<>();
