+++ app/src/main/java/io/github/sspanak/tt9/ime/modes/predictions/WordPredictions.java
@@ -7,8 +7,8 @@ import io.github.sspanak.tt9.ime.helpers.TextField;
 import io.github.sspanak.tt9.languages.EmojiLanguage;
 import io.github.sspanak.tt9.languages.Language;
 import io.github.sspanak.tt9.preferences.settings.SettingsStore;
-import io.github.sspanak.tt9.util.Characters;
 import io.github.sspanak.tt9.util.TextTools;
+import io.github.sspanak.tt9.util.chars.Characters;
 
 public class WordPredictions extends Predictions {
 	private final TextField textField;
